requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking requirements.in
#
asgiref==3.7.2
    # via django
django==4.1
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-filter
    #   django-model-utils
    #   django-polymorphic
    #   djangorestframework
    #   djangorestframework-jsonapi
    #   djangorestframework-simplejwt
django-colorfield==0.8.0
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filter==23.2
    # via djangorestframework-jsonapi
django-model-utils==4.3.1
    # via -r requirements.in
django-polymorphic==3.1.0
    # via djangorestframework-jsonapi
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-jsonapi
    #   djangorestframework-simplejwt
djangorestframework-jsonapi[django-filter,django-polymorphic,openapi]==6.0.0
    # via -r requirements.in
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
factory-boy==3.2.1
    # via -r requirements.in
faker==0.7.4
    # via factory-boy
inflection==0.5.1
    # via djangorestframework-jsonapi
pillow==9.5.0
    # via
    #   -r requirements.in
    #   django-colorfield
psycopg2-binary==2.9.6
    # via -r requirements.in
pyjwt==2.7.0
    # via djangorestframework-simplejwt
python-dateutil==2.8.2
    # via faker
pytz==2023.3
    # via djangorestframework
pyyaml==6.0
    # via djangorestframework-jsonapi
six==1.16.0
    # via
    #   faker
    #   python-dateutil
sqlparse==0.4.4
    # via django
typing-extensions==4.6.3
    # via asgiref
uritemplate==4.1.1
    # via djangorestframework-jsonapi
